refactor: フォルダ名の統一とVRTタグの最適化
- accordion を Accordion にリネーム（ContactFormと統一）
- app/page.tsx のimportパスを更新
- FormField と FormButton のストーリーから vrt タグを削除
  - 細かいコンポーネントは VRT 不要
  - ContactForm 全体のストーリーで VRT を実施
- すべてのテスト成功を確認 (29/29)
